+++ TravelTipsAPI/ViewModels/db_basic/AttractionPatchViewModel.cs
@@ -5,7 +5,7 @@
         public string? Name { get; set; }
         public string? Description { get; set; }
         public string? Address { get; set; }
-        public int? OsmId { get; set; }
+        public long? OsmId { get; set; }
         public int? LinkId { get; set; }
     }
 }
