+++ TravelTipsAPI/ViewModels/db_basic/PreferRoutePatchViewModel.cs
@@ -3,8 +3,8 @@
     public class PreferRoutePatchViewModel
     {
         public int? Type { get; set; }
-        public int? DepartOsmId { get; set; }
-        public int? ArrivalOsmId { get; set; }
+        public long? DepartOsmId { get; set; }
+        public long? ArrivalOsmId { get; set; }
         public int? EstimateTime { get; set; }
         public int? LinkId { get; set; }
     }
